quicklaunch: WIP try to add setup script during install from app loader
